Move build-scripts/profile-stats.py to ssg/_build_profile.py
